# backend/openapi.yaml
openapi: 3.0.0
info:
  title: RYC - Rate Your Courses API
  version: 1.0.0
  description: Privacy-first anonymous course review platform
  contact:
    name: RYC Team
servers:
  - url: http://localhost:3000
    description: Development server
paths:
  /healthz:
    get:
      summary: Health check endpoint
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  timestamp:
                    type: string
                    format: date-time
                  service:
                    type: string
                  database:
                    type: string
  /api/courses:
    get:
      summary: List all courses
      tags:
        - Courses
      parameters:
        - name: department
          in: query
          schema:
            type: string
          description: Filter by department
        - name: search
          in: query
          schema:
            type: string
          description: Search courses by name, code, or description
      responses:
        '200':
          description: List of courses with statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseWithStats'
  /api/courses/{id}:
    get:
      summary: Get course details with reviews
      tags:
        - Courses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Course details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        '404':
          description: Course not found
  /api/reviews:
    post:
      summary: Submit a new review
      tags:
        - Reviews
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
                  status:
                    type: string
        '400':
          description: Invalid input
        '404':
          description: Course not found
  /api/reviews/{id}/flag:
    post:
      summary: Flag a review for moderation
      tags:
        - Reviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required:
                - reason
      responses:
        '200':
          description: Review flagged successfully
        '400':
          description: Invalid input
        '404':
          description: Review not found
components:
  schemas:
    CourseWithStats:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        department:
          type: string
        description:
          type: string
        reviewCount:
          type: integer
        avgRating:
          type: number
          nullable: true
        avgDifficulty:
          type: number
          nullable: true
        avgWorkload:
          type: number
          nullable: true
    CourseDetail:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        department:
          type: string
        description:
          type: string
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
    Review:
      type: object
      properties:
        id:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
        workload:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        semester:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required:
        - courseId
        - rating
        - difficulty
        - workload
      properties:
        courseId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
        workload:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        semester:
          type: string
        year:
          type: integer